Troubleshoot tests

+++ migrations/010.do.create_products.sql
@@ -6,7 +6,7 @@ CREATE TABLE products (
     product_url TEXT NOT NULL,
     feature_image_url TEXT NOT NULL,
     multiple_color_options BOOLEAN NOT NULL,
-    cost_in_home_currency MONEY NOT NULL,
+    cost_in_home_currency double precision NOT NULL,
     wash_id INTEGER REFERENCES wash_instructions(id) ON DELETE CASCADE,
     dry_id INTEGER REFERENCES dry_instructions(id) ON DELETE CASCADE,
     cmt_country TEXT,
